[MIG] l10n_es_pos_by_device: Migration to 16.0
