+++ l10n_es_pos_by_device/README.rst
@@ -7,7 +7,7 @@ Punto de venta adaptado a la legislación española por dispositivo
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:01ebed04f2276a3c527353f7d167a29c45e376b01977b3f6e66889b408ed9edb
+   !! source digest: sha256:f65c9fd543f9562d8038976c9de6a3a77c9f78d1076f35cc64f6fcd09d52933c
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -17,13 +17,13 @@ Punto de venta adaptado a la legislación española por dispositivo
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fl10n--spain-lightgray.png?logo=github
-    :target: https://github.com/OCA/l10n-spain/tree/14.0/l10n_es_pos_by_device
+    :target: https://github.com/OCA/l10n-spain/tree/16.0/l10n_es_pos_by_device
     :alt: OCA/l10n-spain
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/l10n-spain-14-0/l10n-spain-14-0-l10n_es_pos_by_device
+    :target: https://translation.odoo-community.org/projects/l10n-spain-16-0/l10n-spain-16-0-l10n_es_pos_by_device
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/l10n-spain&target_branch=14.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/l10n-spain&target_branch=16.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
@@ -39,14 +39,18 @@ Punto de venta adaptado a la legislación española por dispositivo
 Configuration
 =============
 
-**Activación secuencia por dispositivo**
+**Activación de secuencia por dispositivo**
 
-* Esta configuración se activa en *Ajustes > Punto de venta > Secuenciación*.
-* Acceder a Punto de venta > Configuración > Dispositivos físicos TPV
+* Esta configuración se activa en *Punto de venta > Configuración > Ajustes > Secuenciación > Secuencia TPV por dispositivo*.
+
+**Configuración de dispositivos físicos por TPV**
+
+* Acceder a *Punto de venta > Configuración > Dispositivos físicos TPV*
    - Crear dispositivos y sus secuencias. Cada dispositivo debe tener su
      propia secuencia.
-* Acceder a la configuración del punto de venta
-   - Establecer los dispositivos disponibles en ese punto de venta.
+
+* En *Punto de venta > Configuración > Ajustes > Secuenciación > Dispositivos físicos TPV*.
+   - Establecer los dispositivos físicos disponibles en ese punto de venta.
      Si se deja en blanco, se podrán seleccionar todos los dispositivos creados.
 
 Usage
@@ -75,7 +79,7 @@ pedidos en la sesión.
 
 En caso de pérdida de conexión + incidente en dispositivo (pérdida de batería,
 cierre de ventana o navegador): El dispositivo se queda bloqueado. Tiene que
-ser desbloqueado manualmente para volver a usarse. En caso entrar en la misma
+ser desbloqueado manualmente para volver a usarse. En caso de entrar en la misma
 sesión, los pedidos pendientes se incluirán en la misma sesión. En caso de ser
 una sesión diferente, serán incluidos en una sesión de rescate siguiendo el
 estándar de Odoo.
@@ -86,7 +90,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/l10n-spain/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/l10n-spain/issues/new?body=module:%20l10n_es_pos_by_device%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/l10n-spain/issues/new?body=module:%20l10n_es_pos_by_device%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -105,6 +109,10 @@ Contributors
 
   * Aritz Olea <ao@landoo.es>
   * Gelo Joga <gj@landoo.es>
+
+* `Factor Libre <https://www.factorlibre.com>`_:
+
+  * Juan Carlos Bonilla <juancarlos.bonilla@factorlibre.com>
 
 Maintainers
 ~~~~~~~~~~~
@@ -127,6 +135,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-ao-landoo| 
 
-This module is part of the `OCA/l10n-spain <https://github.com/OCA/l10n-spain/tree/14.0/l10n_es_pos_by_device>`_ project on GitHub.
+This module is part of the `OCA/l10n-spain <https://github.com/OCA/l10n-spain/tree/16.0/l10n_es_pos_by_device>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
