+++ l10n_es_pos_by_device/i18n/es.po
@@ -30,13 +30,14 @@ msgstr ""
 
 #. module: l10n_es_pos_by_device
 #: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_config__pos_device_ids
+#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_config_settings__pos_device_ids
 msgid "Available POS devices"
 msgstr "Dispositivos TPV disponibles"
 
 #. module: l10n_es_pos_by_device
 #. openerp-web
-#: code:addons/l10n_es_pos_by_device/static/src/js/PaymentScreen.js:0
-#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.js:0
+#: code:addons/l10n_es_pos_by_device/static/src/js/PaymentScreen.esm.js:0
+#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.esm.js:0
 #, python-format
 msgid "Cannot establish device. Clossing POS."
 msgstr "No es posible establecer dispositivo. Cerrando TPV."
@@ -72,37 +73,23 @@ msgid "Device locked"
 msgstr "Dispositivo bloqueado"
 
 #. module: l10n_es_pos_by_device
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_config__display_name
 #: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_device__display_name
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_order__display_name
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_session__display_name
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_company__display_name
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_config_settings__display_name
 msgid "Display Name"
 msgstr "Nombre mostrado"
 
 #. module: l10n_es_pos_by_device
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_config__id
 #: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_device__id
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_order__id
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_session__id
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_company__id
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_config_settings__id
 msgid "ID"
-msgstr "ID"
+msgstr ""
 
 #. module: l10n_es_pos_by_device
 #: model:ir.model.fields,help:l10n_es_pos_by_device.field_pos_config__pos_device_ids
+#: model:ir.model.fields,help:l10n_es_pos_by_device.field_res_config_settings__pos_device_ids
 msgid "If left empty, all devices will be selectable"
 msgstr "Si se deja vacío, se pueden seleccionar todos los dispositivos"
 
 #. module: l10n_es_pos_by_device
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_config____last_update
 #: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_device____last_update
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_order____last_update
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_session____last_update
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_company____last_update
-#: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_res_config_settings____last_update
 msgid "Last Modified on"
 msgstr "Última modificación el"
 
@@ -127,8 +114,8 @@ msgid "Name"
 msgstr "Nombre"
 
 #. module: l10n_es_pos_by_device
-#. openerp-web
-#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/DeviceName.js:0
+#. odoo-javascript
+#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/DeviceName.esm.js:0
 #, python-format
 msgid "No device"
 msgstr "Sin dispositivo"
@@ -142,7 +129,7 @@ msgstr "Dispositivo físico TPV"
 #. module: l10n_es_pos_by_device
 #: model:ir.actions.act_window,name:l10n_es_pos_by_device.pos_devices_act_window
 #: model:ir.ui.menu,name:l10n_es_pos_by_device.pos_device_menu
-#: model_terms:ir.ui.view,arch_db:l10n_es_pos_by_device.pos_config_view_form_by_device
+#: model_terms:ir.ui.view,arch_db:l10n_es_pos_by_device.view_res_config_settings_pos_seq_device
 msgid "POS Physical devices"
 msgstr "Dispositivos físicos TPV"
 
@@ -175,17 +162,16 @@ msgid "Secuencing by physical device"
 msgstr "Secuencia TPV por dispositivo"
 
 #. module: l10n_es_pos_by_device
-#. openerp-web
-#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.js:0
+#. odoo-javascript
+#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.esm.js:0
 #, python-format
 msgid "Select Physical Device"
 msgstr "Selecciona dispositivo físico"
 
 #. module: l10n_es_pos_by_device
-#: model_terms:ir.ui.view,arch_db:l10n_es_pos_by_device.pos_config_view_form_by_device
+#: model_terms:ir.ui.view,arch_db:l10n_es_pos_by_device.view_res_config_settings_pos_seq_device
 msgid "Select available physical devices for this POS configuration"
-msgstr ""
-"Selecciona los dispositivos físicos disponibles para esta configuración TPV"
+msgstr "Selecciona los dispositivos físicos disponibles para esta configuración TPV"
 
 #. module: l10n_es_pos_by_device
 #: model:ir.model.fields,field_description:l10n_es_pos_by_device.field_pos_device__sequence
@@ -213,15 +199,16 @@ msgid "Simplified Invoice prefix"
 msgstr "Prefijo de Factura Simplificada"
 
 #. module: l10n_es_pos_by_device
+#. odoo-python
 #: code:addons/l10n_es_pos_by_device/models/pos_config.py:0
 #, python-format
 msgid "There are no physical devices available. Cannot start session."
 msgstr "No hay dispositivos físicos disponibles. No es posible iniciar sesión."
 
 #. module: l10n_es_pos_by_device
-#. openerp-web
-#: code:addons/l10n_es_pos_by_device/static/src/js/PaymentScreen.js:0
-#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.js:0
+#. odoo-javascript
+#: code:addons/l10n_es_pos_by_device/static/src/js/PaymentScreen.esm.js:0
+#: code:addons/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.esm.js:0
 #, python-format
 msgid "There was a connection error when trying to establish the device."
 msgstr "Ha habido un error de conexión al intentar establecer el dispositivo."
@@ -234,7 +221,7 @@ msgstr "Desbloquear el dispositivo"
 #. module: l10n_es_pos_by_device
 #: model_terms:ir.ui.view,arch_db:l10n_es_pos_by_device.view_res_config_settings_pos_seq_device
 msgid "Use a different subsequence for each physical device on POS."
-msgstr "Utilice una subsecuencia diferente para cada dispositivo físico en POS."
+msgstr "Utilice una subsecuencia diferente para cada dispositivo físico en el TPV."
 
 #. module: l10n_es_pos_by_device
 #: model:ir.model,name:l10n_es_pos_by_device.model_pos_device
