+++ l10n_es_pos_by_device/models/__init__.py
@@ -3,3 +3,4 @@ from . import pos_device
 from . import pos_order
 from . import res_company
 from . import res_config_settings
+from . import pos_session
