+++ l10n_es_pos_by_device/models/res_config_settings.py
@@ -9,3 +9,6 @@ class ResConfigSettings(models.TransientModel):
     pos_sequence_by_device = fields.Boolean(
         related="company_id.pos_sequence_by_device", readonly=False
     )
+    pos_device_ids = fields.Many2many(
+        related="pos_config_id.pos_device_ids", readonly=False
+    )
