+++ l10n_es_pos_by_device/readme/CONFIGURE.rst
@@ -1,9 +1,13 @@
-**Activación secuencia por dispositivo**
+**Activación de secuencia por dispositivo**
 
-* Esta configuración se activa en *Ajustes > Punto de venta > Secuenciación*.
-* Acceder a Punto de venta > Configuración > Dispositivos físicos TPV
+* Esta configuración se activa en *Punto de venta > Configuración > Ajustes > Secuenciación > Secuencia TPV por dispositivo*.
+
+**Configuración de dispositivos físicos por TPV**
+
+* Acceder a *Punto de venta > Configuración > Dispositivos físicos TPV*
    - Crear dispositivos y sus secuencias. Cada dispositivo debe tener su
      propia secuencia.
-* Acceder a la configuración del punto de venta
-   - Establecer los dispositivos disponibles en ese punto de venta.
+
+* En *Punto de venta > Configuración > Ajustes > Secuenciación > Dispositivos físicos TPV*.
+   - Establecer los dispositivos físicos disponibles en ese punto de venta.
      Si se deja en blanco, se podrán seleccionar todos los dispositivos creados.
