+++ l10n_es_pos_by_device/readme/CONTRIBUTORS.rst
@@ -2,3 +2,7 @@
 
   * Aritz Olea <ao@landoo.es>
   * Gelo Joga <gj@landoo.es>
+
+* `Factor Libre <https://www.factorlibre.com>`_:
+
+  * Juan Carlos Bonilla <juancarlos.bonilla@factorlibre.com>
