+++ l10n_es_pos_by_device/static/description/index.html
@@ -367,9 +367,9 @@ ul.auto-toc {
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!! source digest: sha256:01ebed04f2276a3c527353f7d167a29c45e376b01977b3f6e66889b408ed9edb
+!! source digest: sha256:f65c9fd543f9562d8038976c9de6a3a77c9f78d1076f35cc64f6fcd09d52933c
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/l10n-spain/tree/14.0/l10n_es_pos_by_device"><img alt="OCA/l10n-spain" src="https://img.shields.io/badge/github-OCA%2Fl10n--spain-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/l10n-spain-14-0/l10n-spain-14-0-l10n_es_pos_by_device"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/l10n-spain&amp;target_branch=14.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
+<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/l10n-spain/tree/16.0/l10n_es_pos_by_device"><img alt="OCA/l10n-spain" src="https://img.shields.io/badge/github-OCA%2Fl10n--spain-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/l10n-spain-16-0/l10n-spain-16-0-l10n_es_pos_by_device"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/l10n-spain&amp;target_branch=16.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
 <ul class="simple">
 <li>Cubre el requerimiento legal de secuencias de facturas simplificadas del TPV
 en el caso de que los dispositivos se queden sin conexión.</li>
@@ -390,11 +390,14 @@ en el caso de que los dispositivos se queden sin conexión.</li>
 </div>
 <div class="section" id="configuration">
 <h1><a class="toc-backref" href="#toc-entry-1">Configuration</a></h1>
-<p><strong>Activación secuencia por dispositivo</strong></p>
+<p><strong>Activación de secuencia por dispositivo</strong></p>
 <ul class="simple">
-<li>Esta configuración se activa en <em>Ajustes &gt; Punto de venta &gt; Secuenciación</em>.</li>
+<li>Esta configuración se activa en <em>Punto de venta &gt; Configuración &gt; Ajustes &gt; Secuenciación &gt; Secuencia TPV por dispositivo</em>.</li>
+</ul>
+<p><strong>Configuración de dispositivos físicos por TPV</strong></p>
+<ul class="simple">
 <li><dl class="first docutils">
-<dt>Acceder a Punto de venta &gt; Configuración &gt; Dispositivos físicos TPV</dt>
+<dt>Acceder a <em>Punto de venta &gt; Configuración &gt; Dispositivos físicos TPV</em></dt>
 <dd><ul class="first last">
 <li>Crear dispositivos y sus secuencias. Cada dispositivo debe tener su
 propia secuencia.</li>
@@ -403,9 +406,9 @@ propia secuencia.</li>
 </dl>
 </li>
 <li><dl class="first docutils">
-<dt>Acceder a la configuración del punto de venta</dt>
+<dt>En <em>Punto de venta &gt; Configuración &gt; Ajustes &gt; Secuenciación &gt; Dispositivos físicos TPV</em>.</dt>
 <dd><ul class="first last">
-<li>Establecer los dispositivos disponibles en ese punto de venta.
+<li>Establecer los dispositivos físicos disponibles en ese punto de venta.
 Si se deja en blanco, se podrán seleccionar todos los dispositivos creados.</li>
 </ul>
 </dd>
@@ -432,7 +435,7 @@ definición del dispositivo.</p>
 pedidos en la sesión.</p>
 <p>En caso de pérdida de conexión + incidente en dispositivo (pérdida de batería,
 cierre de ventana o navegador): El dispositivo se queda bloqueado. Tiene que
-ser desbloqueado manualmente para volver a usarse. En caso entrar en la misma
+ser desbloqueado manualmente para volver a usarse. En caso de entrar en la misma
 sesión, los pedidos pendientes se incluirán en la misma sesión. En caso de ser
 una sesión diferente, serán incluidos en una sesión de rescate siguiendo el
 estándar de Odoo.</p>
@@ -442,7 +445,7 @@ estándar de Odoo.</p>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/l10n-spain/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/l10n-spain/issues/new?body=module:%20l10n_es_pos_by_device%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/l10n-spain/issues/new?body=module:%20l10n_es_pos_by_device%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -461,6 +464,10 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Gelo Joga &lt;<a class="reference external" href="mailto:gj&#64;landoo.es">gj&#64;landoo.es</a>&gt;</li>
 </ul>
 </li>
+<li><a class="reference external" href="https://www.factorlibre.com">Factor Libre</a>:<ul>
+<li>Juan Carlos Bonilla &lt;<a class="reference external" href="mailto:juancarlos.bonilla&#64;factorlibre.com">juancarlos.bonilla&#64;factorlibre.com</a>&gt;</li>
+</ul>
+</li>
 </ul>
 </div>
 <div class="section" id="maintainers">
@@ -472,7 +479,7 @@ mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
 <p>Current <a class="reference external" href="https://odoo-community.org/page/maintainer-role">maintainer</a>:</p>
 <p><a class="reference external image-reference" href="https://github.com/ao-landoo"><img alt="ao-landoo" src="https://github.com/ao-landoo.png?size=40px" /></a></p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/l10n-spain/tree/14.0/l10n_es_pos_by_device">OCA/l10n-spain</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/l10n-spain/tree/16.0/l10n_es_pos_by_device">OCA/l10n-spain</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
