--- l10n_es_pos_by_device/views/pos_templates.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template id="assets" inherit_id="point_of_sale.assets">
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/l10n_es_pos_by_device/static/src/js/models.js"
-            />
-            <script
-                type="text/javascript"
-                src="/l10n_es_pos_by_device/static/src/js/PaymentScreen.js"
-            />
-            <script
-                type="text/javascript"
-                src="/l10n_es_pos_by_device/static/src/js/Screens/Chrome/Chrome.js"
-            />
-            <script
-                type="text/javascript"
-                src="/l10n_es_pos_by_device/static/src/js/Screens/Chrome/DeviceName.js"
-            />
-        </xpath>
-    </template>
-</odoo>
